refactor: InputParser 클래스를 서비스로 이동
