+++ src/main/java/racingcar/controller/RacingGameController.java
@@ -1,9 +1,8 @@
 package racingcar.controller;
 
 import racingcar.domain.car.Car;
-import racingcar.parser.InputParser;
+import racingcar.service.CarSetupService;
 import racingcar.service.RacingGameService;
-import racingcar.validator.InputValidator;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
@@ -13,26 +12,26 @@ public class RacingGameController {
     private final RacingGameService racingGameService;
     private final InputView inputView;
     private final OutputView outputView;
-    private final InputParser inputParser;
+    private final CarSetupService carSetupService;
 
     public RacingGameController(
             RacingGameService racingGameService,
             InputView inputView,
             OutputView outputView,
-            InputParser inputParser) {
+            CarSetupService carSetupService) {
         this.racingGameService = racingGameService;
         this.inputView = inputView;
         this.outputView = outputView;
-        this.inputParser = inputParser;
+        this.carSetupService = carSetupService;
     }
 
     public void run() {
         outputView.displayCarNameInput();
         String inputValue = inputView.inputCarNames();
-        List<Car> CarNames = inputParser.parseCarNames(inputValue);
+        List<Car> CarNames = carSetupService.parseCarNames(inputValue);
 
         outputView.displayAttemptCount();
-        int count = inputParser.parseAttemptCount(inputView.inputAttemptCount());
+        int count = carSetupService.parseAttemptCount(inputView.inputAttemptCount());
     }
